Integration tests for password manager with some refactoring (#15)
This pull request introduces several changes across multiple files,
focusing on code consistency, functionality enhancements, and
refactoring. Key updates include standardizing code style, improving
Dropbox integration, enhancing password management features, and
refining Cypress tests.

### Code Style Consistency:
* Updated all imports and object definitions to use single quotes
instead of double quotes for consistency (`cypress.config.ts`,
`cypress/support/component.ts`, `src/app/layout.tsx`,
`src/components/app/Home.tsx`).
[[1]](diffhunk://#diff-3f62e885350a48939f240b9daf40f6b06d41da7751f2b613238fc7669ac0a1b2L1-R7)
[[2]](diffhunk://#diff-5ab71f4254fac269ededf6e398bf8bae6ae00d9ca962820b3524bc9b3a6a784eL17-R22)
[[3]](diffhunk://#diff-29d0a55f5ca1a9951042587f8b891bb4cd8b204f5bc855dc2468ed44cde2f5b3L22-L31)
[[4]](diffhunk://#diff-169d66fe226ce0dcd6630dcc9d734a56f50c6b502d1a0d0aff3320765bb5eebeL83-R92)

### Password Management Improvements:
* Added a new "Updated" sorting option to the `PasswordTable` component
and ensured case-insensitive filtering for entries
(`src/components/dashboard/PasswordManager/PasswordTable.tsx`).
[[1]](diffhunk://#diff-12b328dde93609a1f0449c9db9d74dba61f26dc0634b934c84f70aeeb1e57035R33)
[[2]](diffhunk://#diff-12b328dde93609a1f0449c9db9d74dba61f26dc0634b934c84f70aeeb1e57035L172-R193)
[[3]](diffhunk://#diff-12b328dde93609a1f0449c9db9d74dba61f26dc0634b934c84f70aeeb1e57035R208-R221)
* Enhanced the `PasswordTable` component’s dropdown menus and buttons
with better formatting and additional attributes for testing
(`src/components/dashboard/PasswordManager/PasswordTable.tsx`).
[[1]](diffhunk://#diff-12b328dde93609a1f0449c9db9d74dba61f26dc0634b934c84f70aeeb1e57035R245)
[[2]](diffhunk://#diff-12b328dde93609a1f0449c9db9d74dba61f26dc0634b934c84f70aeeb1e57035R255)

### Refactoring and Cleanup:
* Removed unused imports and streamlined logic for better readability
and maintainability (`src/app/page.tsx`, `src/components/app/Home.tsx`).
[[1]](diffhunk://#diff-73d7a23e5015801b9bcc9db601d6ec9594d3eb34e5bb23154e0ae4b0c30f1a3bL2-R18)
[[2]](diffhunk://#diff-169d66fe226ce0dcd6630dcc9d734a56f50c6b502d1a0d0aff3320765bb5eebeL43-R52)
* Simplified the `RootLayout` component by merging nested JSX elements
and removing redundant cleanup code (`src/app/layout.tsx`).
[[1]](diffhunk://#diff-29d0a55f5ca1a9951042587f8b891bb4cd8b204f5bc855dc2468ed44cde2f5b3L45-R44)
[[2]](diffhunk://#diff-29d0a55f5ca1a9951042587f8b891bb4cd8b204f5bc855dc2468ed44cde2f5b3L71-R74)

### Cypress Testing Improvements:
* Refactored the `HomePageWrapper` component in Cypress tests to use
dependency injection, enabling more flexible test setups
(`src/components/app/Home.cy.tsx`).
* Enhanced test cases to verify Dropbox OAuth flows and user
interactions with mock services (`src/components/app/Home.cy.tsx`).

+++ src/app/page.tsx
@@ -1,21 +1,22 @@
 'use client';
-import { useEffect, useRef, useState } from 'react';
-import { connectDropbox, getAuthUrl, hasRedirectedFromAuth } from 'lib/dropbox';
+import { useEffect } from 'react';
+import { getAuthUrl } from 'lib/dropbox';
 import Home from 'components/app/Home';
-import { useUser, useUserDispatch } from 'contexts/user.context';
-import { UserStatus } from 'contexts/reducers/user.reducer';
+import { useUserDispatch } from 'contexts/user.context';
 import { useRouter } from 'next/navigation';
 import { Routes, useLocation } from 'contexts/location.context';
-import { APP_URL } from 'lib/constants';
+import { APP_URL, DROPBOX_CLIENT_ID } from 'lib/constants';
 import { useTrezorUiEvents } from 'hooks/use-trezor-ui-events';
 import { useTrezorDeviceEvents } from 'hooks/use-trezor-device-events';
-import { DropboxSessionStatus, useDropboxSession } from 'hooks/use-dropbox-session';
+import { useDropboxSession } from 'hooks/use-dropbox-session';
+import { useDropboxWindowOauthParams } from 'hooks/use-dropbox-window-oauth-params';
 export default function App() {
   const router = useRouter();
   const [location, _] = useLocation();
   const [userDispatch] = useUserDispatch();
-  const dropboxStatus = useDropboxSession();
-// Link Trezor events to user context
+  const { search, codeVerifier } = useDropboxWindowOauthParams();
+  const dropboxStatus = useDropboxSession(search, DROPBOX_CLIENT_ID, codeVerifier, );
+  // Link Trezor events to user context
   useTrezorUiEvents();
   // Link Trezor device events to user context
   useTrezorDeviceEvents();
@@ -27,17 +28,23 @@ export default function App() {
   }, [location, router]);
 
   const handleDropBoxSignIn = () => {
+    if (DROPBOX_CLIENT_ID === undefined) {
+      console.error('DROPBOX_CLIENT_ID is undefined');
+      return;
+    }
     if (APP_URL === undefined) {
       console.error('APP_URI is undefined');
       return;
     }
-    getAuthUrl(APP_URL).then(({authUrl, codeVerifier}) => {
-      window.sessionStorage.clear();
-      window.sessionStorage.setItem('codeVerifier', codeVerifier);
-      window.location.href = authUrl as string;
-    }).catch((error) => {
-      console.error(error);
-    });
+    getAuthUrl(APP_URL, DROPBOX_CLIENT_ID)
+      .then(({ authUrl, codeVerifier }) => {
+        window.sessionStorage.clear();
+        window.sessionStorage.setItem('codeVerifier', codeVerifier);
+        window.location.href = authUrl as string;
+      })
+      .catch((error) => {
+        console.error(error);
+      });
   };
 
   const handleLogout = () => {
@@ -47,9 +54,12 @@ export default function App() {
 
   return (
     <Home
-      initialLoadingStatus={dropboxStatus !== DropboxSessionStatus.NOT_CONNECTED}
+      dropboxArgs={{
+        urlSearch: search,
+        codeVerifier: codeVerifier,
+      }}
       handleDropBoxSignIn={handleDropBoxSignIn}
-      handleLogout={handleLogout}>
-    </Home>
+      handleLogout={handleLogout}
+    ></Home>
   );
 }
